+++ main.go
@@ -25,6 +25,9 @@ var BuildDate = "Unknown"
 // HTTP connection params
 const defaultHTTPHost = "127.0.0.1"
 const defaultHTTPPort = 8825
+const defaultHTTPReadTimeout = 1
+const defaultHTTPWriteTimeout = 1
+const defaultHTTPIdleTimeout = 1
 
 // StatsD connection params
 const defaultStatsDHost = "127.0.0.1"
@@ -33,6 +36,9 @@ const defaultStatsDPort = 8125
 // declare command line options
 var httpHost = flag.String("http-host", defaultHTTPHost, "HTTP Host")
 var httpPort = flag.Int("http-port", defaultHTTPPort, "HTTP Port")
+var httpReadTimeout = flag.Int("http-timeout-read", defaultHTTPReadTimeout, "The maximum duration in seconds for reading the entire request, including the body")
+var httpWriteTimeout = flag.Int("http-timeout-write", defaultHTTPWriteTimeout, "The maximum duration in seconds before timing out writes of the respons")
+var httpIdleTimeout = flag.Int("http-timeout-idle", defaultHTTPIdleTimeout, "The maximum amount of time in seconds to wait for the next request when keep-alives are enabled")
 var tlsCert = flag.String("tls-cert", "", "TLS certificate to enable HTTPS")
 var tlsKey = flag.String("tls-key", "", "TLS private key  to enable HTTPS")
 var statsdHost = flag.String("statsd-host", defaultStatsDHost, "StatsD Host")
@@ -72,6 +78,9 @@ func main() {
 	proxyServer := proxy.NewServer(
 		*httpHost,
 		*httpPort,
+		*httpReadTimeout,
+		*httpWriteTimeout,
+		*httpIdleTimeout,
 		*statsdHost,
 		*statsdPort,
 		*tlsCert,